dco-backend/requirements.txt
Flask 
Flask-RESTful 
Flask-SQLAlchemy 
flask-swagger-ui
psycopg2-binary
flask_cors
ngrok
invoke